+++ Spec/Programs/ChatServer.xs
@@ -42,7 +42,7 @@ class Room {
     }
 
     fn Broadcast(message: Message) {
-        for ((_, sender): (ClientId, Thread.sender<Message>) in this.members) {
+        for ((else, sender): (ClientId, Thread.sender<Message>) in this.members) {
             sender.send(message);
         }
     }