+++ tests/fixtures/source/RecursiveCallMain.xs
@@ -1,4 +1,6 @@
 // SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
 // SPDX-License-Identifier: Apache-2.0
 
-fn main() -> Long { return (1, 2); }
+fn main() -> Long {
+    return main();
+}